fixed smartbench name for drywall
