+++ src/asmcnc/apps/systemTools_app/screens/screen_factory_settings.py
@@ -33,8 +33,7 @@ from asmcnc.apps.systemTools_app.screens.calibration import screen_set_threshold
 from asmcnc.apps.systemTools_app.screens.calibration import screen_general_measurement
 from asmcnc.production.database.calibration_database import CalibrationDatabase
 
-from asmcnc.comms.model_manager import ModelManagerSingleton
-from asmcnc.comms.router_machine import ProductCodes
+from asmcnc.comms.model_manager import ModelManagerSingleton, ProductCodes
 
 Builder.load_string(
     """
